+++ Gambo.ECS/EcsContextBuilder.cs
@@ -2,6 +2,9 @@
 
 namespace Gambo.ECS
 {
+    /// <summary>
+    ///     Builder class for EcsContext
+    /// </summary>
     public class EcsContextBuilder
     {
         private readonly EcsContext m_context;
@@ -12,10 +15,10 @@ namespace Gambo.ECS
         }
 
         /// <summary>
-        /// Adds an existing registry to the context.
+        ///     Adds an existing registry to the context.
         /// </summary>
         /// <param name="registry">The registry to attach</param>
-        /// <returns></returns>
+        /// <returns>The builder</returns>
         public EcsContextBuilder WithRegistry(EcsRegistry registry)
         {
             m_context.Registry = registry;
@@ -24,10 +27,10 @@ namespace Gambo.ECS
         }
 
         /// <summary>
-        /// Attaches a service provider to the context. Used to resolve services that may be specified in a system.
+        ///     Attaches a service provider to the context. Used to resolve services that may be specified in a system.
         /// </summary>
         /// <param name="serviceProvider">The IServiceProvider to use for service resolution.</param>
-        /// <returns></returns>
+        /// <returns>The builder</returns>
         public EcsContextBuilder WithServiceProvider(IServiceProvider serviceProvider)
         {
             m_context.ServiceProvider = serviceProvider;
@@ -36,9 +39,9 @@ namespace Gambo.ECS
         }
 
         /// <summary>
-        /// Builds the context.
+        ///     Builds the context.
         /// </summary>
-        /// <returns></returns>
+        /// <returns>The resulting EcsContext of the build pipeline</returns>
         public EcsContext Build()
         {
             return m_context;