+++ Gambo.ECS/EcsRegistry.cs
@@ -5,6 +5,9 @@ using System.Linq;
 
 namespace Gambo.ECS
 {
+    /// <summary>
+    ///     Represents a collection of entities and their attached components
+    /// </summary>
     public class EcsRegistry
     {
         private static int m_nextRegistryId;
@@ -23,17 +26,42 @@ namespace Gambo.ECS
             m_components = new Dictionary<EcsEntity, List<object>>();
         }
 
+        /// <summary>
+        ///     The amount of entities contained in the registry
+        /// </summary>
         public int EntitiesCount => m_entities.Count;
 
+        /// <summary>
+        ///     A dictionary of the entity-component relationships contained in the registry
+        /// </summary>
         public ReadOnlyDictionary<EcsEntity, List<object>> Components =>
             new(m_components);
 
+        /// <summary>
+        ///     Called whenever an entity is added to the registry
+        /// </summary>
         public event EventHandler<EntityEventArgs>? OnEntityAdded;
+        
+        /// <summary>
+        ///     Called whenever an entity is removed from the registry
+        /// </summary>
         public event EventHandler<EntityEventArgs>? OnEntityRemoved;
 
+        /// <summary>
+        ///     Called whenever a component is attached to an entity
+        /// </summary>
         public event EventHandler<ComponentEventArgs>? OnComponentAdded;
+        
+        /// <summary>
+        ///     Called whenever a component is removed from an entity
+        /// </summary>
         public event EventHandler<ComponentEventArgs>? OnComponentRemoved;
 
+        /// <summary>
+        ///     Fetches the entity with the specified Id
+        /// </summary>
+        /// <param name="entityID">The id of the entity</param>
+        /// <returns>The requested entity if it exists in the registry, null if not</returns>
         public EcsEntity? GetEntity(int entityID)
         {
             foreach (var entity in m_entities)
@@ -43,6 +71,10 @@ namespace Gambo.ECS
             return null;
         }
 
+        /// <summary>
+        ///     Creates a new entity and attaches it to the registry
+        /// </summary>
+        /// <returns>The new entity</returns>
         public EcsEntity CreateEntity()
         {
             var entity = new EcsEntity(m_nextEntityId, this);
@@ -54,15 +86,12 @@ namespace Gambo.ECS
             return entity;
         }
 
-        public bool AddEntity(EcsEntity entity)
-        {
-            var added = m_entities.Add(entity);
-
-            if (added) OnEntityAdded?.Invoke(this, new EntityEventArgs(entity));
-
-            return added;
-        }
-
+        /// <summary>
+        ///     Removes the specified entity from the registry
+        /// </summary>
+        /// <param name="entity">Entity to remove</param>
+        /// <param name="permanent">Set to true to remove all attached components</param>
+        /// <returns>True if removed, false if not found</returns>
         public bool RemoveEntity(EcsEntity entity, bool permanent = false)
         {
             var removed = m_entities.Remove(entity);
@@ -74,6 +103,14 @@ namespace Gambo.ECS
             return removed;
         }
 
+        /// <summary>
+        ///     Attaches a new component of the specified type to specified entity
+        /// </summary>
+        /// <param name="componentType">Component type to attach</param>
+        /// <param name="entity">Entity to attach component to</param>
+        /// <param name="args">Component constructor arguments</param>
+        /// <returns>The created component</returns>
+        /// <exception cref="ArgumentException">Thrown if the component is not a struct, or if the type has no compatible constructor</exception>
         public object AddComponent(Type componentType, EcsEntity entity, params object[] args)
         {
             AssertEntity(entity);
@@ -107,23 +144,32 @@ namespace Gambo.ECS
             return component;
         }
 
+        /// <summary>
+        ///     Removes the specified component from the entity
+        /// </summary>
+        /// <param name="entity">The entity to remove the component from</param>
+        /// <typeparam name="T">The type of component to remove</typeparam>
         public void RemoveComponent<T>(EcsEntity entity) where T : struct
         {
             if (!m_components.ContainsKey(entity)) return;
 
             var entityComponents = m_components[entity];
-            for (var i = 0; i < entityComponents.Count; i++)
+            for (int i = 0; i < entityComponents.Count; i++)
             {
-                var component = entityComponents.ElementAt(i);
-                if (component.GetType() == typeof(T))
-                {
-                    entityComponents.Remove(component);
-                    OnComponentRemoved?.Invoke(this, new ComponentEventArgs(entity, component));
-                    break;
-                }
+                object component = entityComponents.ElementAt(i);
+                if (component.GetType() != typeof(T)) continue;
+                
+                entityComponents.Remove(component);
+                OnComponentRemoved?.Invoke(this, new ComponentEventArgs(entity, component));
+                break;
             }
         }
 
+        /// <summary>
+        ///     Gets all components attached to the specified entity
+        /// </summary>
+        /// <param name="entity">The entity to query for components</param>
+        /// <returns>A list of all components attached to the entity</returns>
         public List<object> GetComponents(EcsEntity entity)
         {
             AssertEntity(entity);
@@ -131,6 +177,11 @@ namespace Gambo.ECS
             return m_components.ContainsKey(entity) ? m_components[entity] : new List<object>();
         }
 
+        /// <summary>
+        ///     Gets all components of a specified type in the registry
+        /// </summary>
+        /// <typeparam name="TComponent">The type of components to query</typeparam>
+        /// <returns>All components of the given type present in the registry</returns>
         public IEnumerable<TComponent> GetComponentsOfType<TComponent>() where TComponent : struct
         {
             var componentsByType = m_components
@@ -142,6 +193,12 @@ namespace Gambo.ECS
             return componentsByType;
         }
 
+        /// <summary>
+        ///     Gets a specific type of component on an entity
+        /// </summary>
+        /// <param name="entity">The entity to query</param>
+        /// <typeparam name="T">The type of components to query</typeparam>
+        /// <returns>The first matching component found</returns>
         public T? GetComponent<T>(EcsEntity entity) where T : struct
         {
             AssertEntity(entity);
@@ -156,6 +213,11 @@ namespace Gambo.ECS
             return null;
         }
 
+        /// <summary>
+        ///     Replaces the instance of a component type with the specified instance
+        /// </summary>
+        /// <param name="component">The component instance to replace. Type will be inferred from the instance</param>
+        /// <param name="entity">The entity to manipulate</param>
         public void ReplaceComponent(object component, EcsEntity entity)
         {
             AssertEntity(entity);